requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
passlib>=1.7.4
bcrypt<4.0
python-multipart
python-jose
psycopg2-binary
python-dotenv
fastapi-mail
email-validator
